Add short tagged enums

+++ src/main/java/moe/nea/jdwp/primitives/JDWPByte.kt
@@ -6,16 +6,16 @@ import moe.nea.jdwp.JDWPReader
 import moe.nea.jdwp.JDWPSingleState
 import moe.nea.jdwp.JDWPWriter
 
-class JDWPByte() : JDWPSingleState<Byte>() {
-    constructor(value: Byte) : this() {
+class JDWPByte() : JDWPSingleState<UByte>() {
+    constructor(value: UByte) : this() {
         this.value = value
     }
 
     override fun read(reader: JDWPReader) {
-        value = reader.consume(1)[0].toByte()
+        value = reader.consume(1)[0]
     }
 
     override fun write(writer: JDWPWriter) {
-        writer.append(ubyteArrayOf(value!!.toUByte()))
+        writer.append(ubyteArrayOf(value!!))
     }
 }
